rapihin (refactor) - joe
buat keperluan intent, biar lebih readable
