--- app/src/main/java/com/u_binusportal/Category.java
@@ -1,42 +1,0 @@
-package com.u_binusportal;
-
-import android.graphics.drawable.Drawable;
-import android.net.Uri;
-
-import com.pchmn.materialchips.model.ChipInterface;
-
-public class Category implements ChipInterface {
-
-    private Object categoryId;
-    private String categoryName;
-
-    public Category(String categoryId, String categoryName) {
-        this.categoryId = categoryId;
-        this.categoryName = categoryName;
-    }
-
-    @Override
-    public Object getId() {
-        return categoryId;
-    }
-
-    @Override
-    public Uri getAvatarUri() {
-        return null;
-    }
-
-    @Override
-    public Drawable getAvatarDrawable() {
-        return null;
-    }
-
-    @Override
-    public String getLabel() {
-        return categoryName;
-    }
-
-    @Override
-    public String getInfo() {
-        return null;
-    }
-}
