--- app/src/main/java/com/u_binusportal/SlidePagerAdapter.java
@@ -1,32 +1,0 @@
-package com.u_binusportal;
-
-import android.annotation.SuppressLint;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentStatePagerAdapter;
-
-import java.util.List;
-
-public class SlidePagerAdapter extends FragmentStatePagerAdapter {
-
-    private FragmentManager fm;
-    private List<Fragment> fragmentList;
-
-    public SlidePagerAdapter(@NonNull FragmentManager fm, int behaviour, List<Fragment> fragmentList) {
-        super(fm, behaviour);
-        this.fragmentList = fragmentList;
-    }
-
-    @NonNull
-    @Override
-    public Fragment getItem(int position) {
-        return fragmentList.get(position);
-    }
-
-    @Override
-    public int getCount() {
-        return fragmentList.size();
-    }
-}
