+++ app/src/main/java/com/u_binusportal/activity/pageFragmentsActivities/pageFragment/PageThree.java
@@ -1,0 +1,23 @@
+package com.u_binusportal.activity.pageFragmentsActivities.pageFragment;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.u_binusportal.R;
+
+public class PageThree extends Fragment {
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.awal_3, container, false);
+        return rootView;
+    }
+}
