--- app/src/main/java/com/u_binusportal/fragment/Profil/ProfilModel.java
@@ -1,17 +1,0 @@
-package com.u_binusportal.fragment.Profil;
-
-import android.os.Bundle;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.u_binusportal.R;
-
-public class ProfilModel extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.pencarian);
-    }
-}
