+++ app/src/main/java/com/u_binusportal/handlers/FragmentHandler.java
@@ -1,0 +1,35 @@
+package com.u_binusportal.handlers;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+import androidx.navigation.ui.AppBarConfiguration;
+import androidx.navigation.ui.NavigationUI;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+import com.u_binusportal.R;
+
+public class FragmentHandler extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setContentView(R.layout.group_fragment); // main page
+        BottomNavigationView botNav = findViewById(R.id.bottom_menu_navigation); // set bar di bawah
+
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
+
+        AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
+            R.id.navigation_home, R.id.navigation_search, R.id.navigation_profil).build();
+        NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
+
+        NavigationUI.setupWithNavController(botNav, navController);
+    }
+}
+
