--- app/src/main/java/com/u_binusportal/pageFragments/PageTwo.java
@@ -1,23 +1,0 @@
-package com.u_binusportal.pageFragments;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.u_binusportal.R;
-
-public class PageTwo extends Fragment {
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             @Nullable ViewGroup container,
-                             @Nullable Bundle savedInstanceState) {
-        ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.awal_2, container, false);
-        return rootView;
-    }
-}
